requirements.txt
arab-transcription
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
autoflake==2.1.1
autopep8==2.0.2
black==23.3.0
Brotli==1.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cryptography==41.0.2
cssselect2==0.7.0
Django==4.1.3
django-cors-headers==3.14.0
django-filter==22.1
django-restframework==0.0.1
django-weasyprint==2.2.0
djangorestframework==3.14.0
drf-nested-routers==0.93.4
flake8==6.0.0
fonttools==4.41.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
lxml==4.9.3
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==23.1
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.7.0
pyflakes==3.0.1
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
pypdf==3.13.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
pytz==2022.6
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzlocal==5.0.1
uritemplate==4.1.1
uritools==4.0.1
urllib3==2.0.4
weasyprint==58.0
webencodings==0.5.1
xhtml2pdf==0.2.11
zipp==3.10.0
zopfli==0.2.2
